+++ src/index.js
@@ -16,13 +16,13 @@ import * as serviceWorker from "./serviceWorker";
 import config from "./custom/config";
 import { isIphoneAndCordova } from "./utils";
 import { gtagInit } from "./gtag.js";
-import LocationProvider, { useGPSLocation } from "./LocationProvider";
-import { usePhotos } from "./PhotosProvider";
-import { useSelectedFeature } from "./SelectedFeatureProvider";
-import { useConfig } from "./ConfigProvider";
-import { useOnline } from "./OnlineProvider";
-import UserProvider, { useUser } from "./UserProvider";
-import { useStats } from "./StatsProvider";
+import LocationProvider, { useGPSLocation } from "./providers/LocationProvider";
+import { usePhotos } from "./providers/PhotosProvider";
+import { useSelectedFeature } from "./providers/SelectedFeatureProvider";
+import { useConfig } from "./providers/ConfigProvider";
+import { useOnline } from "./providers/OnlineProvider";
+import UserProvider, { useUser } from "./providers/UserProvider";
+import { useStats } from "./providers/StatsProvider";
 import * as Sentry from "@sentry/browser";
 import { dbFirebase } from "features/firebase";
