+++ src/providers/PhotosProvider.tsx
@@ -13,7 +13,7 @@ import { Location, History } from "history";
 import { RealtimeUpdate } from "features/firebase/dbFirebase";
 import _ from "lodash";
 import useAsyncEffect from "hooks/useAsyncEffect";
-import { useUser } from "UserProvider";
+import { useUser } from "./UserProvider";
 
 const photoToFeature = (photo: Photo): Feature => ({
   feature: "Feature",
